libflux: create reactor_private header file
Problem: We wish to move some reactor functions out of libflux, but there
are some shared structs / functions that will be needed if we do so.

Solution: Create a reactor_private.h header file that can be used internally by
flux-core utility libraries.  Move struct flux_reactor, struct_watcher,
events_to_libev, and libev_to_events into the header.
